Moved gradle plugin out of sample app and added some additional methods for OttreliteAndroid
