+++ plugin/ottrelite-gradle-plugin/ottrelite-plugin/src/main/java/com/callstack/plugin/OttreliteInstrumentation.kt
@@ -4,12 +4,12 @@ import com.android.build.api.instrumentation.AsmClassVisitorFactory
 import com.android.build.api.instrumentation.ClassContext
 import com.android.build.api.instrumentation.ClassData
 import com.android.build.api.instrumentation.InstrumentationParameters
+import com.callstack.plugin.model.replacementDescriptorList
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Input
 import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.MethodVisitor
 import org.objectweb.asm.Opcodes
-
 
 interface ReactInstrumentationParameters : InstrumentationParameters {
 
@@ -31,11 +31,8 @@ abstract class OttreliteClassVisitorFactory :
         classContext: ClassContext,
         nextClassVisitor: ClassVisitor
     ): ClassVisitor {
-        // We return a custom ClassVisitor that will do the actual work.
-        // It wraps the 'nextClassVisitor' in the chain.
         return object : ClassVisitor(Opcodes.ASM9, nextClassVisitor) {
 
-            // This method is called for each method in the class.
             override fun visitMethod(
                 access: Int,
                 name: String?,
@@ -45,7 +42,6 @@ abstract class OttreliteClassVisitorFactory :
             ): MethodVisitor {
                 val mv = super.visitMethod(access, name, descriptor, signature, exceptions)
 
-                // Check if this method matches our configured target.
                 val className = classContext.currentClassData.className
                 val desiredClassFound = className == parameters.get().targetClassName.get()
                 val replacementDescriptor = if (desiredClassFound) {
@@ -55,10 +51,9 @@ abstract class OttreliteClassVisitorFactory :
                         replacementDescriptor != null &&
                         (access and Opcodes.ACC_STATIC) != 0
 
-                if (isTargetMethod && replacementDescriptor != null) {
-                    println("[OttreliteInstrumentation] Modifying method: $className.$name$descriptor")
+                if (isTargetMethod) {
+                    println("Modifying method: $className.$name$descriptor")
 
-                    // Return a new MethodVisitor that will replace the method's body.
                     return replacementDescriptor.methodVisitorType.getMethodVisitor(
                         mv,
                         replacementDescriptor.additionalMethodName, access
@@ -72,56 +67,5 @@ abstract class OttreliteClassVisitorFactory :
     override fun isInstrumentable(classData: ClassData): Boolean {
         return classData.className == parameters.get().targetClassName.get()
     }
-}
-
-private val replacementDescriptorList = buildList {
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "isTracing",
-            targetMethodDescriptor = "(J)Z",
-            additionalMethodName = null,
-            methodVisitorType = MethodVisitorType.BOOLEAN
-        )
-    )
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "endSection",
-            targetMethodDescriptor = "(J)V",
-            additionalMethodName = "endEvent",
-            methodVisitorType = MethodVisitorType.LONG
-        )
-    )
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "beginSection",
-            targetMethodDescriptor = "(JLjava/lang/String;)V",
-            additionalMethodName = "beginEvent",
-            methodVisitorType = MethodVisitorType.LONG_STRING
-        )
-    )
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "beginAsyncSection",
-            targetMethodDescriptor = "(JLjava/lang/String;I)V",
-            additionalMethodName = "beginAsyncEvent",
-            methodVisitorType = MethodVisitorType.LONG_STRING_INT
-        )
-    )
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "endAsyncSection",
-            targetMethodDescriptor = "(JLjava/lang/String;I)V",
-            additionalMethodName = "endAsyncEvent",
-            methodVisitorType = MethodVisitorType.LONG_STRING_INT
-        )
-    )
-    add(
-        ReplacementDescriptor(
-            targetMethodName = "traceCounter",
-            targetMethodDescriptor = "(JLjava/lang/String;I)V",
-            additionalMethodName = "traceCounter",
-            methodVisitorType = MethodVisitorType.LONG_STRING_INT
-        )
-    )
 }
