+++ plugin/ottrelite-gradle-plugin/settings.gradle.kts
@@ -1,0 +1,3 @@
+rootProject.name = "ottrelite-gradle-plugin"
+
+include("ottrelite-plugin")
